fix : service, repository 리팩토링
