+++ src/main/java/sideproject/puddy/service/DogService.java
@@ -13,11 +13,8 @@ import sideproject.puddy.dto.tag.TagDto;
 import sideproject.puddy.exception.CustomException;
 import sideproject.puddy.exception.ErrorCode;
 import sideproject.puddy.model.Dog;
-import sideproject.puddy.model.DogTagMap;
 import sideproject.puddy.model.Person;
 import sideproject.puddy.repository.DogRepository;
-import sideproject.puddy.repository.DogTagMapRepository;
-import sideproject.puddy.repository.RegisterNumberRepository;
 import sideproject.puddy.security.util.SecurityUtil;
 
 import java.time.LocalDate;
@@ -28,13 +25,10 @@ import java.util.List;
 @RequiredArgsConstructor
 public class DogService {
     private final DogRepository dogRepository;
-    private final DogTagService dogTagService;
     private final DogTypeService dogTypeService;
-    private final DogTagMapRepository dogTagMapRepository;
+    private final DogTagMapService dogTagMapService;
     private final AuthService authService;
     private final DogTransactionalService dogTransactionalService;
-    private final RegisterNumberRepository registerNumberRepository;
-
     public ResponseEntity<String> saveDogBySignUp(Long personId, PostDogRequest request){
         Person person = authService.findById(personId);
         dogTransactionalService.save(person, request);
@@ -49,8 +43,8 @@ public class DogService {
     public ResponseEntity<String> updateDog(Long id, UpdateDogRequest request){
         Dog dog = findByPersonAndId(id);
         dog.updateDog(request.getImage(), dogTypeService.findByContent(request.getType()), request.isGender(), request.isNeuter());
-        dogTagMapRepository.deleteAll(dogTagMapRepository.findAllByDog(dog));
-        request.getTags().forEach(dogTag -> dogTagMapRepository.save(new DogTagMap(dog, dogTagService.findByContent(dogTag.getContent()))));
+        dogTagMapService.deleteAllTags(dog);
+        dogTagMapService.saveAllTags(request.getTags(), dog);
         return ResponseEntity.ok().body("ok");
     }
     @Transactional
@@ -63,9 +57,6 @@ public class DogService {
         Dog dog = findByPersonAndId(id);
         List<TagDto> tags = dog.getDogTagMaps().stream().map(dogTagMap -> new TagDto(dogTagMap.getDogTag().getContent())).toList();
         return new DogDetailResponse(dog.getImage(), dog.isGender(), dog.getDogType().getContent(), dog.isNeuter(), tags);
-    }
-    public boolean existRegisterNum(Long registerNum){
-        return registerNumberRepository.existsByRegisterNum(registerNum);
     }
     public Dog findByPersonAndId(Long id){
         Person person = authService.findById(SecurityUtil.getCurrentUserId());
