+++ src/main/java/sideproject/puddy/service/DogTransactionalService.java
@@ -6,10 +6,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import sideproject.puddy.dto.dog.request.PostDogRequest;
 import sideproject.puddy.model.Dog;
-import sideproject.puddy.model.DogTagMap;
 import sideproject.puddy.model.Person;
 import sideproject.puddy.repository.DogRepository;
-import sideproject.puddy.repository.DogTagMapRepository;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -18,14 +16,13 @@ import java.time.format.DateTimeFormatter;
 @RequiredArgsConstructor
 public class DogTransactionalService {
     private final DogRepository dogRepository;
-    private final DogTagService dogTagService;
-    private final DogTagMapRepository dogTagMapRepository;
+    private final DogTagMapService dogTagMapService;
     private final DogTypeService dogTypeService;
     @Transactional
     public void save(Person person, PostDogRequest request){
         boolean main = !dogRepository.existsByPerson(person);
         Dog dog = dogRepository.save(new Dog(request.getName(), request.isGender(), LocalDate.parse(request.getBirth(), DateTimeFormatter.ISO_DATE)
                 , dogTypeService.findByContent(request.getType()), request.getRegisterNum(), request.isNeuter(), main, request.getImage(), person));
-        request.getTags().forEach(dogTag -> dogTagMapRepository.save(new DogTagMap(dog, dogTagService.findByContent(dogTag.getContent()))));
+        dogTagMapService.saveAllTags(request.getTags(), dog);
     }
 }
