+++ src/main/java/sideproject/puddy/service/RegisterNumberService.java
@@ -1,0 +1,16 @@
+package sideproject.puddy.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import sideproject.puddy.repository.RegisterNumberRepository;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+public class RegisterNumberService {
+    private final RegisterNumberRepository registerNumberRepository;
+    public boolean existRegisterNum(Long registerNum){
+        return registerNumberRepository.existsByRegisterNum(registerNum);
+    }
+}
